Refatorações. Início do Lexador em EguaP.
