+++ fontes/avaliador-sintatico/index.ts
@@ -1,4 +1,4 @@
-import tiposDeSimbolos from '../tiposDeSimbolos';
+import tiposDeSimbolos from '../lexador/tipos-de-simbolos';
 import { performance } from 'perf_hooks';
 import { AvaliadorSintaticoInterface, SimboloInterface } from '../interfaces';
 import {
