+++ fontes/interpretador/dialetos/egua-classico.ts
@@ -1,4 +1,4 @@
-import tiposDeSimbolos from '../../tiposDeSimbolos';
+import tiposDeSimbolos from '../../lexador/tipos-de-simbolos';
 import { Ambiente } from '../../ambiente';
 import { Delegua } from '../../delegua';
 import carregarBibliotecaGlobal from '../../bibliotecas/biblioteca-global';
