+++ fontes/interpretador/index.ts
@@ -2,7 +2,7 @@ import * as caminho from 'path';
 import * as fs from 'fs';
 import { performance } from 'perf_hooks';
 
-import tiposDeSimbolos from '../tiposDeSimbolos';
+import tiposDeSimbolos from '../lexador/tipos-de-simbolos';
 
 import { Ambiente } from '../ambiente';
 import { Delegua } from '../delegua';
