+++ fontes/lexador/dialetos/index.ts
@@ -1,1 +1,1 @@
-export * from './egua-classico'
+export * from './lexador-egua-classico'
