+++ fontes/lexador/dialetos/tipos-de-simbolos-eguap.ts
@@ -1,0 +1,71 @@
+export default {
+    ADICAO: "ADICAO",
+    BIT_AND: "BIT_AND",
+    BIT_OR: "BIT_OR",
+    BIT_XOR: "BIT_XOR",
+    BIT_NOT: "BIT_NOT",
+    CASO: "CASO",
+    CHAVE_DIREITA: "CHAVE_DIREITA",
+    CHAVE_ESQUERDA: "CHAVE_ESQUERDA",
+    CLASSE: "CLASSE",
+    COLCHETE_DIREITO: "COLCHETE_DIREITO",
+    COLCHETE_ESQUERDO: "COLCHETE_ESQUERDO",
+    CONTINUA: "CONTINUA",
+    DIFERENTE: "DIFERENTE",
+    DIVISAO: "DIVISAO",
+    DOIS_PONTOS: "DOIS_PONTOS",
+    E: "E",
+    EM: "EM",
+    ENQUANTO: "ENQUANTO",
+    EOF: "EOF",
+    ESCOLHA: "ESCOLHA",
+    ESCREVA: "ESCREVA",
+    ESPACO_INDENTACAO: "ESPACO_INDENTACAO",
+    EXPONENCIACAO: "EXPONENCIACAO",
+    IGUAL: "IGUAL",
+    IGUAL_IGUAL: "IGUAL_IGUAL",
+    FALSO: "FALSO",
+    FAZER: "FAZER",
+    FINALMENTE: "FINALMENTE",
+    FUNCAO: "FUNCAO",
+    FUNÇÃO: "FUNÇÃO",
+    HERDA: "HERDA",
+    IDENTIFICADOR: "IDENTIFICADOR",
+    IMPORTAR: "IMPORTAR",
+    ISTO: "ISTO",
+    MAIOR: "MAIOR",
+    MAIOR_IGUAL: "MAIOR_IGUAL",
+    MAIOR_MAIOR: "MAIOR_MAIOR",
+    MAIS_IGUAL: "MAIS_IGUAL",
+    MENOR: "MENOR",
+    MENOS_IGUAL: "MENOS_IGUAL",
+    MENOR_IGUAL: "MENOR_IGUAL",
+    MENOR_MENOR: "MENOR_MENOR",
+    MODULO: "MODULO",
+    MULTIPLICACAO: "MULTIPLICACAO",
+    NEGACAO: "NEGACAO",
+    NULO: "NULO",
+    NUMERO: "NUMERO",
+    OU: "OU",
+    PADRAO: "PADRAO",
+    PARA: "PARA",
+    PARENTESE_DIREITO: "PARENTESE_DIREITO",
+    PARENTESE_ESQUERDO: "PARENTESE_ESQUERDO",
+    PAUSA: "PAUSA",
+    PEGUE: "PEGUE",
+    PONTO: "PONTO",
+    PONTO_E_VIRGULA: "PONTO_E_VIRGULA",
+    RETORNA: "RETORNA",
+    SUBTRACAO: "SUBTRACAO",
+    SE: "SE",
+    SENAO: "SENAO",
+    SENÃO: "SENÃO",
+    SENAOSE: "SENAOSE",
+    SENÃOSE: "SENÃOSE",
+    SUPER: "SUPER",
+    TENTE: "TENTE",
+    TEXTO: "TEXTO",
+    VARIAVEL: "VARIAVEL",
+    VERDADEIRO: "VERDADEIRO",
+    VIRGULA: "VIRGULA"
+};
