+++ fontes/lexador/tipos-de-simbolos.ts
